[edit] 예약 상세 페이지에 계좌 정보 입력 모달 추가 및 예약 취소 로직 수정

+++ src/components/account/ReservationDetail.tsx
@@ -11,6 +11,11 @@ import Modal from "../Modal";
 import { useShowModal } from "@/hooks/useShowModal";
 import { CloseButton } from "../controls/Button";
 import { showToast } from "@/utils/toast";
+import { JoinInput } from "@/components/controls/Inputs";
+import { JoinSelect } from "@/components/controls/Select";
+import { bankCodeList } from "@/lib/bankCodeList";
+import { validations } from "@/utils/validations";
+
 interface DetailDataProps {
   label: string;
   data: string;
@@ -21,7 +26,14 @@ export default function ReservationDetail({ bookId }: { bookId: string }) {
     null
   );
   const { showModal, handleShowModal } = useShowModal();
+  const [modalType, setModalType] = useState<"qr" | "account" | null>(null);
+  const [selectAccount, setSelectAccount] = useState("000");
+  const [accountNum, setAccountNum] = useState("");
+  const [depositor, setDepositor] = useState("");
   const router = useRouter();
+  const isAccountNumInputValid =
+    validations.accountNum.safeParse(accountNum).success;
+
   useEffect(() => {
     async function fetchDetailData() {
       try {
@@ -35,18 +47,25 @@ export default function ReservationDetail({ bookId }: { bookId: string }) {
     fetchDetailData();
   }, []);
 
-  const handleCancelBooking = async (bookId: string) => {
-    try {
-      const response = await axios.delete(`/api/account/book/${bookId}`);
-      if (response.status === 200) {
-        showToast(
-          "예매 취소가 완료되었습니다. 마이페이지로 이동합니다.",
-          "success",
-          () => router.push("/account")
-        );
+  const handleCancelBooking = async (status: string, bookId: string) => {
+    if (status === "booked") {
+      showToast("환불 계좌 입력 후 취소를 진행해주세요");
+      handleShowModal(true);
+      setModalType("account");
+      return;
+    } else {
+      try {
+        const response = await axios.delete(`/api/account/book/${bookId}`);
+        if (response.status === 200) {
+          showToast(
+            "예매 취소가 완료되었습니다. 마이페이지로 이동합니다.",
+            "success",
+            () => router.push("/account")
+          );
+        }
+      } catch (error) {
+        console.error("Error fetching booking details:", error);
       }
-    } catch (error) {
-      console.error("Error fetching booking details:", error);
     }
   };
 
@@ -57,6 +76,7 @@ export default function ReservationDetail({ bookId }: { bookId: string }) {
     //   handleShowModal(true);
     // }
     handleShowModal(true);
+    setModalType("qr");
   };
 
   const handleModalClose = () => {
@@ -81,7 +101,9 @@ export default function ReservationDetail({ bookId }: { bookId: string }) {
             <ul className="absolute top-0 right-0 text-xs flex gap-2">
               <li>
                 <Button
-                  onClick={(e) => handleCancelBooking(bookId)}
+                  onClick={(e) =>
+                    handleCancelBooking(detailData.paymentStatus, bookId)
+                  }
                   className="font-normal py-[2.5px] sm:text-base sm:font-bold"
                   disabled={isPerformanceEnded as boolean}
                 >
@@ -196,7 +218,7 @@ export default function ReservationDetail({ bookId }: { bookId: string }) {
           </section>
           {/* QR 확인 모달 */}
           <Modal
-            isOpen={showModal}
+            isOpen={showModal && modalType === "qr"}
             onClose={() => handleModalClose()}
             className="relative p-[0px]"
           >
@@ -206,6 +228,57 @@ export default function ReservationDetail({ bookId }: { bookId: string }) {
                 className="absolute top-6 right-6"
                 onClick={() => handleModalClose()}
               />
+            </>
+          </Modal>
+          {/* 계좌정보 입력 모달 */}
+          <Modal
+            isOpen={showModal && modalType === "account"}
+            onClose={() => handleModalClose()}
+            className="relative p-8"
+          >
+            <>
+              <form>
+                <CloseButton
+                  className="absolute top-4 right-4"
+                  onClick={() => handleModalClose()}
+                />
+                <JoinSelect
+                  label="은행정보"
+                  value={selectAccount}
+                  onChange={setSelectAccount}
+                  className="mb-5"
+                  options={bankCodeList}
+                />
+                <JoinInput
+                  label="계좌번호"
+                  value={accountNum}
+                  onChange={setAccountNum}
+                  validation={validations.accountNum}
+                  placeholder="숫자 (‘-’ 문자 제외)"
+                  max={14}
+                />
+                <JoinInput
+                  label="예금주"
+                  value={depositor}
+                  onChange={setDepositor}
+                  placeholder="10자 이내"
+                  validation={validations.depositor}
+                  max={10}
+                />
+                <Button
+                  className="float-right py-2 px-4"
+                  size="small"
+                  highlight
+                  disabled={
+                    selectAccount === "000" ||
+                    !isAccountNumInputValid ||
+                    !depositor
+                  }
+                  type="submit"
+                >
+                  환불요청
+                </Button>
+              </form>
             </>
           </Modal>
         </>
